add part4 step19-21; oh jwt!

+++ part4/controllers/blogs.js
@@ -1,5 +1,6 @@
 const blogsRouter = require('express').Router()
 const jwt = require('jsonwebtoken')
+const blog = require('../models/blog')
 
 const Blog = require('../models/blog')
 const User = require('../models/user')
@@ -30,7 +31,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (req, response) => {
-  result = await Blog.findByIdAndDelete(req.params.id)
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+  if (!req.token || !decodedToken.id)
+    return response.status(401).json({ error: 'token missing or invalid' })
+
+  const user = await User.findById(decodedToken.id)
+
+  const blog = await Blog.findById(req.params.id)
+  if (blog.user.toString() !== user.id.toString())
+    return response.status(403).json({error: 'you are not allowed to do that!'})
+
+  const result = await Blog.findByIdAndDelete(req.params.id)
   return response.status(204).end()
 })
